Solution to Calculate Average Marks of Students

+++ Calculate Average Marks of Students@@67877d61dc107486ad5c2743/code.c
@@ -15,7 +15,7 @@ int main(){
         scanf("%f",&s[i].marks);
         temp = temp + s[i].marks;
     }
-       int average = temp / n;
+       float average = temp / n;
         printf("Average Marks: %.2f",average);
 
 }
